rename Router to FormRouter and update imports. Move BrowserRouter to examples index
