+++ examples/burial-form/index.jsx
@@ -1,5 +1,5 @@
-import React, { useContext } from 'react';
-import {Navigate, Route, useLocation} from 'react-router-dom'
+import React from 'react';
+import { BrowserRouter, Route} from 'react-router-dom'
 import {FormRouter, ConditionalRoute} from '@department-of-veterans-affairs/va-forms-system-core';
 import BurialIntroduction from './BurialIntroduction';
 import ClaimantInformation from './ClaimantInformation';
@@ -31,33 +31,35 @@ const BurialApp = (props) => {
 
   return (
     <div className='vads-u-display--flex vads-u-align-items--center vads-u-flex-direction--column'>
-      <FormRouter basename={props.basename}
-        formData={props.initialValues}
-        title="Burials Example"
-        transformForSubmit={mapProps}
+      <BrowserRouter basename={props.basename}>
+        <FormRouter
+          formData={props.initialValues}
+          title="Burials Example"
+          transformForSubmit={mapProps}
         >
-        <Route index element={<BurialIntroduction title="Introduction Page" />} />
-        <Route path="/claimant-information" element={<ClaimantInformation title="Claimant Information" />} />
-        <Route path="/veteran-information" element={<VeteranInformation title="Deceased Veteran Information" />} />
-        <Route path="/veteran-information/burial" element={<BurialInformation title="Deceased Veteran Information" />} />
-        <Route path="/military-history/service-periods" element={<MilitaryServiceHistory title="Military Service History" />} />
-        <Route path="/military-history/previous-names" element={<PreviousNames title="Military history" />} />
-        <Route path="/benefits/selection" element={<BenefitsSelection title="Benefits Selection" />} />
-        <Route path="/benefits/burial-allowance" element={
-          <ConditionalRoute title="Benefits Selection" type="conditional" condition={'benefitsSelection.burialAllowance'}>
-            <BurialAllowance />
-          </ConditionalRoute>}
-        />
-        <Route path="/benefits/plot-allowance" element={
-          <ConditionalRoute title="Benefits Selection" type="conditional" condition={'benefitsSelection.plotAllowance'}>
-            <PlotAllowance />
-          </ConditionalRoute>}
-        />
-        <Route path="/claimant-contact-information" element={<ClaimantContactInformation title="Claimant contact information" />} />
-        <Route path="/review-and-submit" element={<ReviewPage title="Review Your Application" />} />
-        <Route path="/confirmation" element={<ConfirmationPage title="Claim submitted" />} />
-        <Route path="*" element={<NoMatch name="No Routes for App" />} />
-      </FormRouter>
+          <Route index element={<BurialIntroduction title="Introduction Page" />} />
+          <Route path="/claimant-information" element={<ClaimantInformation title="Claimant Information" />} />
+          <Route path="/veteran-information" element={<VeteranInformation title="Deceased Veteran Information" />} />
+          <Route path="/veteran-information/burial" element={<BurialInformation title="Deceased Veteran Information" />} />
+          <Route path="/military-history/service-periods" element={<MilitaryServiceHistory title="Military Service History" />} />
+          <Route path="/military-history/previous-names" element={<PreviousNames title="Military history" />} />
+          <Route path="/benefits/selection" element={<BenefitsSelection title="Benefits Selection" />} />
+          <Route path="/benefits/burial-allowance" element={
+            <ConditionalRoute title="Benefits Selection" type="conditional" condition={'benefitsSelection.burialAllowance'}>
+              <BurialAllowance />
+            </ConditionalRoute>}
+          />
+          <Route path="/benefits/plot-allowance" element={
+            <ConditionalRoute title="Benefits Selection" type="conditional" condition={'benefitsSelection.plotAllowance'}>
+              <PlotAllowance />
+            </ConditionalRoute>}
+          />
+          <Route path="/claimant-contact-information" element={<ClaimantContactInformation title="Claimant contact information" />} />
+          <Route path="/review-and-submit" element={<ReviewPage title="Review Your Application" />} />
+          <Route path="/confirmation" element={<ConfirmationPage title="Claim submitted" />} />
+          <Route path="*" element={<NoMatch name="No Routes for App" />} />
+        </FormRouter>
+      </BrowserRouter>
     </div>
   )
 }
