+++ examples/chapter-form/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet, Route } from 'react-router-dom'
-import FormRouter from '../../src/routing/Router';
+import FormRouter from '../../src/routing/FormRouter';
 import Chapter from '../../src/routing/Chapter';
 import Page from '../../src/routing/Page';
 import IntroductionPage from '../../src/form-layout/IntroductionPage'
