+++ test/form-layout/FormFooter.test.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Route } from 'react-router-dom';
-import FormRouter from '../../src/routing/Router';
+import FormRouter from '../../src/routing/FormRouter';
 import FormFooter from '../../src/form-layout/FormFooter';
 
-
 describe('FormFooter', () => {
-
   test('Renders footer', () => {
-    const { container } = render(
-      <FormFooter />
-    );
+    const { container } = render(<FormFooter />);
     const footerContainer = container.querySelector('.help-footer-box');
 
     expect(footerContainer).toBeTruthy();
@@ -18,7 +14,7 @@ describe('FormFooter', () => {
 
   test('Renders footer from router', () => {
     const { container } = render(
-      <FormRouter basename="/" title="hello world" formData={{firstName: ''}}>
+      <FormRouter basename="/" title="hello world" formData={{ firstName: '' }}>
         <Route path="/" element={<div>Hello World</div>} />
       </FormRouter>
     );
